+++ apps/client/src/pages/login/index.ts
@@ -1,1 +1,1 @@
-export { default as LoginPage } from './ui/login-page';
+export { default as LoginPage } from './login-page';
