+++ apps/client/src/pages/login/login-page.tsx
@@ -2,12 +2,11 @@ import { Navigate } from 'react-router';
 
 import { Icon } from '@cds/icon';
 
+import { LOGIN_END_POINT } from '@pages/login/apis/end-point';
+import LoginButton from '@pages/login/components/login-button/login-button';
+
 import { PATH } from '@shared/router/path';
 import { getAccessToken } from '@shared/storage/token-storage';
-
-import LoginButton from '@features/login-button/ui/login-button';
-
-import { LOGIN_END_POINT } from '@widgets/login-page/api/end-point';
 
 import * as styles from './login-page.css';
