+++ apps/client/src/shared/router/lazy.ts
@@ -3,7 +3,7 @@ import { lazy } from 'react';
 export const NewMemoPage = lazy(
   () => import('@pages/new-memo/ui/new-memo-page'),
 );
-export const LoginPage = lazy(() => import('@pages/login/ui/login-page'));
+export const LoginPage = lazy(() => import('@pages/login/login-page'));
 export const AllMemoPage = lazy(
   () => import('@pages/all-memo/ui/all-memo-page'),
 );
